test: split slow backend tests into an integration tier
The unit gate caps the package at 120s and it was already taking ~123s
before this branch, so the next test added to it would have failed CI
regardless of what that test was. Two groups own most of the budget and
neither is a unit test: TestInvalidS3Host drives connect retries against a
host that refuses everything (20.7s), and the four dual-open tests each
stand up a second VB engine over one volume to reproduce the daemon plus
nbdkit topology (14.7s).

Gate both behind an `integration` build tag with a make target and a CI
job, so they still run on every push rather than becoming tests nobody
executes. TestInvalidS3Host moves to its own file because the one it lived
in also holds TestMain and thirty other tests. The unit gate drops to
~104s; the read-modify-write reproducer stays in it, costing 0.15s.

The tag is additive, so the integration target compiles the untagged tests
too and carries a budget covering both.
